Refactor coverage into modular backends
